requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==3.2.2
django-allauth==0.48.0
django-cleanup==6.0.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
oauthlib==3.2.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python-docx==0.8.11
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
